data_source_devices: migrate to plugin framework
Fixes tailscale/corp#37225
