+++ tailscale/provider_framework.go
@@ -191,9 +191,10 @@ func (p *tailscaleProvider) DataSources(_ context.Context) []func() datasource.D
 	return []func() datasource.DataSource{
 		New4Via6DataSource,
 		NewACLDataSource,
-		NewDeviceDataSource,
 		NewMultipleUsersDataSource,
 		NewSingleUserDataSource,
+		NewMultipleDevicesDataSource,
+		NewSingleDeviceDataSource,
 	}
 }
